Refactored VPN code changes
- VPN actions collected in class VPNManager
- Actions on isolated network collected in class IsolatedNetworkManager
- Put all docker-related classes in subfolder Docker which also contains small utils file
